Update 2nd exercise
