+++ 05-crud-rest-api/exercises/05-02-todo-create/README.md
@@ -6,11 +6,11 @@ Understand the flow of building REST API with Postman
 
 ## Exercises:
 
-1. Update Postman "Create Todo" to have description.
+1. Update Postman "Create Todo" to have title and description.
 2. Test Postman to see the result.
-3. Add new field for "desc" (description)
-4. Try Postman "Create Todo" to see the result with description.
-5. Check if the Postman "List Todos" have the description.
+3. Implement create todo handler in `app.js`.
+4. Try Postman "Create Todo" to see the result.
+5. Check if the Postman "List Todos" have newly added todo.
 
 Request:
 
@@ -23,7 +23,7 @@ Request Body:
 ```json
 {
   "title": "Shopping for the food",
-  "desc": "at BigC"
+  "description": "at BigC"
 }
 ```
 
@@ -34,7 +34,7 @@ Response:
   "data": {
     "id": 4,
     "title": "Shopping for the food",
-    "desc": "at BigC"
+    "description": "at BigC"
   }
 }
 ```
